mstest: Rewrite unit test
Main target of this commit was to rewrite unit test for module from old MSTest to new MSTest.
This was completed, but after rewrite, memory leak was detected.
This commit fixes also memory leak, which was occured, when user was trying to allocate incorrect string.
Memory was allocated, but when function was trying to insert it, error occured and function was returning NULL without freeing memory.
Issue #23